lib: utils/fdt: Allow zero length mtime when parsing ACLINT
Currently, the fdt_parse_aclint_node() does not allow the first reg to
be zero length. This may cause the parser to fail when the ACLINT does
not support mtime.

To be robust, delay the check of first reg in fdt_parse_aclint_node()
so the parser can handle ACLINT without mtime support.

Signed-off-by: Inochi Amaoto <inochiama@outlook.com>